# server/index.js
const express = require('express');
const path = require('path');
const fs = require('fs');
const cors = require('cors');
const cluster = require('cluster');
const swaggerUi = require('swagger-ui-express');
const numCPUs = require('os').cpus().length;

const PORT = process.env.PORT || 3001;

const { getFilteredMaterials, materialsJson } = require('./rest/materials');
const recipesJson = require('./rest/recipes');
const { getFilteredArmor, armorJson } = require('./rest/armor');
const { getFilteredFood, foodJson } = require('./rest/food');
const { getFilteredMonsters, monstersJson } = require('./rest/monsters');
const { getFilteredShields, shieldsJson } = require('./rest/shields');
const weaponsJson = require('./rest/weapons');
const { getFilteredBows, bowsJson } = require('./rest/bows');
const { getFilteredAnimals, animalsJson } = require('./rest/animals');
const otherJson = require('./rest/other');
const changelogJson = require('./rest/changelog');

// Multi-process to utilize all CPU cores.
if (cluster.isMaster) {
  console.error(`Node cluster master ${process.pid} is running`);

  // Fork workers.
  for (let i = 0; i < numCPUs; i++) {
    cluster.fork();
  }

  cluster.on('exit', (worker, code, signal) => {
    console.error(`Node cluster worker ${worker.process.pid} exited: code ${code}, signal ${signal}`);
  });

} else {
  const app = express();

  // Allow API and Swagger UI to be called from other origins (e.g. Try it out from api-docs)
  app.use(cors());

  // Priority serve any static files.
  app.use(express.static(path.resolve(__dirname, '../react-ui/build')));

  // Swagger / OpenAPI docs (must be before catch-all)
  app.get('/api-docs/openapi.yaml', function (req, res) {
    res.set('Content-Type', 'application/yaml');
    res.send(fs.readFileSync(path.join(__dirname, 'openapi.yaml'), 'utf8'));
  });
  app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(null, {
    swaggerOptions: { url: '/api-docs/openapi.yaml' }
  }));

  //
  const obj = {
    message: "Hello from the custom server!",
    test: {
      one: "nested message"
    }
  }

  // Answer API requests.
  app.get('/api', function (req, res) {
    res.set('Content-Type', 'application/json');
    res.send(JSON.stringify(obj));
  });

  // Materials API
  app.get('/api/materials', function (req, res) {
    res.set('Content-Type', 'application/json');
    res.send(getFilteredMaterials(req.query));
  });
  // Materials Detail API
  app.get('/api/materials/:id', function (req, res) {
    res.set('Content-Type', 'application/json');
    res.send(materialsJson[Number(req.params.id) - 1]);
  });

  // Recipes API
  app.get('/api/recipes', function (req, res) {
    res.set('Content-Type', 'application/json');
    res.send(recipesJson);
  })
  // Recipes Detail API
  app.get('/api/recipes/:id', function (req, res) {
    res.set('Content-Type', 'application/json');
    res.send(recipesJson[Number(req.params.id) - 1]);
  });

  // Armor API (supports filters: defenseMin/Max, sellPriceMin/Max, bodyPart, addedEffect, setBonus, availability)
  app.get('/api/armor', function (req, res) {
    res.set('Content-Type', 'application/json');
    res.send(getFilteredArmor(req.query));
  })
  // Armor Detail API
  app.get('/api/armor/:id', function (req, res) {
    res.set('Content-Type', 'application/json');
    res.send(armorJson[Number(req.params.id) - 1]);
  });

  // Food API (supports filters: sellPriceMin/Max, hpRecoveryMin/Max, ingredient)
  app.get('/api/food', function (req, res) {
    res.set('Content-Type', 'application/json');
    res.send(getFilteredFood(req.query));
  })
  // Food Detail API
  app.get('/api/food/:id', function (req, res) {
    res.set('Content-Type', 'application/json');
    res.send(foodJson[Number(req.params.id) - 1]);
  });

  // Monsters API
  // Monsters API (supports filters: size, hpMin/Max, rankMin/Max, commonLocation, itemDrop)
  app.get('/api/monsters', function (req, res) {
    res.set('Content-Type', 'application/json');
    res.send(getFilteredMonsters(req.query));
  })
  // Monsters Detail API
  app.get('/api/monsters/:id', function (req, res) {
    res.set('Content-Type', 'application/json');
    res.send(monstersJson[Number(req.params.id) - 1]);
  });

  // Shields API
  // Shields API (supports filters: durabilityMin/Max, parryPowerMin/Max, availability)
  app.get('/api/shields', function (req, res) {
    res.set('Content-Type', 'application/json');
    res.send(getFilteredShields(req.query));
  })
  // Shields Detail API
  app.get('/api/shields/:id', function (req, res) {
    res.set('Content-Type', 'application/json');
    res.send(shieldsJson[Number(req.params.id) - 1]);
  });

  // Weapons API
  app.get('/api/weapons', function(req, res) {
    res.set('Content-Type', 'application/json');
    res.send(weaponsJson);
  })
  // Weapons Detail API
  app.get('/api/weapons/:id', function (req, res) {
    res.set('Content-Type', 'application/json');
    res.send(weaponsJson[Number(req.params.id) - 1]);
  });

   // Bows API (supports filters: attackPowerMin/Max, durabilityMin/Max, rangeMin/Max, multipleArrowsMin/Max, quickShot, availability)
  app.get('/api/bows', function (req, res) {
    res.set('Content-Type', 'application/json');
    res.send(getFilteredBows(req.query));
  })
  // Bows Detail API
  app.get('/api/bows/:id', function (req, res) {
    res.set('Content-Type', 'application/json');
    res.send(bowsJson[Number(req.params.id) - 1]);
  });

  // Animals API (supports ?uniqueCookingEffect=, ?commonLocation=, ?recoverableMaterial=)
  app.get('/api/animals', function (req, res) {
    res.set('Content-Type', 'application/json');
    if (req.query)
      res.send(getFilteredAnimals(req.query));
    else
      res.send(animalsJson)
  });
  // Animals Detail API
  app.get('/api/animals/:id', function (req, res) {
    res.set('Content-Type', 'application/json');
    res.send(animalsJson[Number(req.params.id) - 1]);
  });

  // Other items API
  app.get('/api/other', function (req, res) {
    res.set('Content-Type', 'application/json');
    res.send(otherJson);
  });
  // Other Detail API
  app.get('/api/other/:id', function (req, res) {
    res.set('Content-Type', 'application/json');
    res.send(otherJson[Number(req.params.id) - 1]);
  });

  // Changelog API (static snapshot)
  app.get('/api/changelog', function (req, res) {
    res.set('Content-Type', 'application/json');
    res.send(changelogJson);
  });

  // Ensure unknown API routes never fall back to index.html
  app.get('/api/*', function (req, res) {
    res.status(404).json({ error: 'Unknown API route' });
  });

  // All remaining requests return the React app, so it can handle routing.
  app.get('*', function(request, response) {
    response.sendFile(path.resolve(__dirname, '../react-ui/build', 'index.html'));
  });

  app.listen(PORT, function () {
    console.error(`Node cluster worker ${process.pid}: listening on port ${PORT}`);
  });
}