# server/openapi.yaml
openapi: 3.0.3
info:
  title: Slatepedia API
  description: |
    Slatepedia exposes a REST API over encyclopedia data for *The Legend of Zelda: Breath of the Wild*. All responses are JSON.

    **Resources**

    - **Materials** – Ingredients, plants, ores, and other collectibles (names, types, descriptions, sell prices, locations, recipes that use them).
    - **Recipes** – Cooked dishes and their materials, effects, and bonus effects.
    - **Armor** – Armor pieces with defense, set names, and upgrade levels.
    - **Food** – Cooked food items and their effects.
    - **Monsters** – Enemies and creatures, with descriptions and item drops (weapons, materials, other).
    - **Shields**, **Weapons**, **Bows** – Equipment with attack/defense and metadata.
    - **Animals** – Wildlife (non-enemy) with common locations, recoverable materials, and unique cooking effects. List endpoint supports optional query filters: `uniqueCookingEffect`, `commonLocation`, `recoverableMaterial` (comma-separated for multiple values).
    - **Other** – Miscellaneous items (rupees, key items, bridles, saddles, compendium categories, etc.).
    - **Changelog** – Static list of changelog entries (semantic `version`, date, title, message).

    **Usage**

    Use the **Servers** dropdown to choose the base URL (e.g. production). For each operation, click **Try it out**, fill any parameters, then **Execute** to send a request and see the response. List endpoints return arrays; detail endpoints use a numeric path parameter (e.g. `/api/materials/1`).
  version: 1.0.0

servers:
  - url: /api
    description: Current host (relative – use when docs and API share the same origin)

tags:
  - name: General
    description: Root and changelog
  - name: Materials
  - name: Recipes
  - name: Armor
  - name: Food
  - name: Monsters
  - name: Shields
  - name: Weapons
  - name: Bows
  - name: Animals
  - name: Other
  - name: Changelog

paths:
  /:
    get:
      tags: [General]
      summary: API root
      operationId: getApiRoot
      responses:
        '200':
          description: Simple welcome message
          content:
            application/json:
              schema:
                type: object
                example: { message: "Hello from the custom server!", test: { one: "nested message" } }

  /changelog:
    get:
      tags: [Changelog]
      summary: List changelog entries
      operationId: getChangelog
      description: Returns a static snapshot of changelog entries (version, date, title, message). Versions follow semver-style 0.major.minor with no GitHub commit links.
      responses:
        '200':
          description: Array of changelog entries
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    version: { type: string, example: "0.6.0" }
                    date: { type: string, format: date-time }
                    title: { type: string }
                    message: { type: string }

  /materials:
    get:
      tags: [Materials]
      summary: List materials (with optional filters)
      operationId: getMaterials
      description: |
        Returns materials. Optionally narrow by type, sell price and HP recovery ranges, category name,
        bonus effect (category added effect), potency grade, duration factor range, and availability.
        Multiple values for categorical fields use comma-separated strings (OR semantics within each field).
        Use reserved token "__none__" to match rows with no category, no bonus effect, or no potency grade.
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
          description: Filter by material type (exact match). Comma-separated for multiple.
          example: "Fruit"
        - name: sellPriceMin
          in: query
          required: false
          schema: { type: integer }
          description: Minimum sell price in Rupees (inclusive).
        - name: sellPriceMax
          in: query
          required: false
          schema: { type: integer }
          description: Maximum sell price in Rupees (inclusive).
        - name: hpRecoveryMin
          in: query
          required: false
          schema: { type: integer }
          description: Minimum HP recovery (inclusive).
        - name: hpRecoveryMax
          in: query
          required: false
          schema: { type: integer }
          description: Maximum HP recovery (inclusive).
        - name: category
          in: query
          required: false
          schema:
            type: string
          description: Filter by category name (exact match). Use "__none__" for materials with no category. Comma-separated for multiple.
          example: "Hearty"
        - name: bonusEffect
          in: query
          required: false
          schema:
            type: string
          description: Filter by bonus effect (category added effect, exact match). Use "__none__" when there is no category or no effect. Comma-separated for multiple.
          example: "Extra Hearts"
        - name: potencyGrade
          in: query
          required: false
          schema:
            type: string
          description: Filter by potency grade letter (exact match). Use "__none__" for an empty grade. Comma-separated for multiple.
          example: "A"
        - name: durationFactorMin
          in: query
          required: false
          schema: { type: integer }
          description: Minimum duration factor in seconds (inclusive).
        - name: durationFactorMax
          in: query
          required: false
          schema: { type: integer }
          description: Maximum duration factor in seconds (inclusive).
        - name: availability
          in: query
          required: false
          schema:
            type: string
          description: Filter by exact availability string (matches any element of availabilities[]). Comma-separated for multiple.
          example: "Hyrule Field"
      responses:
        '200':
          description: Array of material objects (filtered if query params provided)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Material' }
  /materials/{id}:
    get:
      tags: [Materials]
      summary: Get material by ID
      operationId: getMaterialById
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Single material object
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Material' }

  /recipes:
    get:
      tags: [Recipes]
      summary: List all recipes
      operationId: getRecipes
      responses:
        '200':
          description: Array of recipe objects
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Recipe' }
  /recipes/{id}:
    get:
      tags: [Recipes]
      summary: Get recipe by ID
      operationId: getRecipeById
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Single recipe object
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Recipe' }

  /armor:
    get:
      tags: [Armor]
      summary: List armor (with optional filters)
      operationId: getArmor
      description: |
        Returns armor. Optionally narrow by numeric ranges (defense, sell price) and/or categories
        (body part, added effect, set bonus, availability). Multiple categorical values use a
        comma-separated string. Empty added effect/set bonus filters use reserved token "__none__"
        as a value name.
      parameters:
        - name: defenseMin
          in: query
          required: false
          schema: { type: integer }
          description: Minimum defense value (inclusive).
        - name: defenseMax
          in: query
          required: false
          schema: { type: integer }
          description: Maximum defense value (inclusive).
        - name: sellPriceMin
          in: query
          required: false
          schema: { type: integer }
          description: Minimum sell price in Rupees (inclusive). Use for items with explicit sell prices.
        - name: sellPriceMax
          in: query
          required: false
          schema: { type: integer }
          description: Maximum sell price in Rupees (inclusive).
        - name: bodyPart
          in: query
          required: false
          schema:
            type: string
          description: Filter by exact body part (e.g. "Head", "Chest"). Comma-separated for multiple matches (OR semantics).
          example: "Chest"
        - name: addedEffect
          in: query
          required: false
          schema:
            type: string
          description: Exact match against added effects. Use "__none__" to match items without an added effect. Comma-separated for multiple values.
          example: "Cold Resistance"
        - name: setBonus
          in: query
          required: false
          schema:
            type: string
          description: Exact match against set bonuses. Use "__none__" for items without a set bonus. Comma-separated for multiple values.
          example: "Attack Up"
        - name: availability
          in: query
          required: false
          schema:
            type: string
          description: Exact match against availability strings. Comma-separated for multiple matches.
          example: "Shop: Tarrey Town"
      responses:
        '200':
          description: Array of armor objects (filtered if query params provided)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Armor' }
  /armor/{id}:
    get:
      tags: [Armor]
      summary: Get armor by ID
      operationId: getArmorById
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Single armor object
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Armor' }

  /food:
    get:
      tags: [Food]
      summary: List food (with optional filters)
      operationId: getFood
      description: |
        Returns food. Optionally narrow by sell price and HP recovery ranges, and/or ingredient.
        Multiple ingredients use a comma-separated string (matches if any selected ingredient appears in ingredients[]).
      parameters:
        - name: sellPriceMin
          in: query
          required: false
          schema: { type: integer }
          description: Minimum sell price in Rupees (inclusive).
        - name: sellPriceMax
          in: query
          required: false
          schema: { type: integer }
          description: Maximum sell price in Rupees (inclusive).
        - name: hpRecoveryMin
          in: query
          required: false
          schema: { type: integer }
          description: Minimum HP recovery (inclusive).
        - name: hpRecoveryMax
          in: query
          required: false
          schema: { type: integer }
          description: Maximum HP recovery (inclusive).
        - name: ingredient
          in: query
          required: false
          schema:
            type: string
          description: Filter by ingredient name (exact match) or ingredient id. Comma-separated for multiple (matches if any listed ingredient is present).
          example: "Hyrule Bass"
      responses:
        '200':
          description: Array of food objects (filtered if query params provided)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Food' }
  /food/{id}:
    get:
      tags: [Food]
      summary: Get food by ID
      operationId: getFoodById
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Single food object
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Food' }

  /monsters:
    get:
      tags: [Monsters]
      summary: List monsters (with optional filters)
      operationId: getMonsters
      description: |
        Returns monsters. Optionally narrow by size, HP and rank ranges, common locations, and item drops.
        Multiple values per categorical field use comma-separated strings (OR semantics within each field).
      parameters:
        - name: size
          in: query
          required: false
          schema:
            type: string
          description: Filter by size (exact match, e.g. "small", "medium", "large"). Comma-separated for multiple.
          example: "large"
        - name: hpMin
          in: query
          required: false
          schema: { type: integer }
          description: Minimum HP (inclusive).
        - name: hpMax
          in: query
          required: false
          schema: { type: integer }
          description: Maximum HP (inclusive).
        - name: rankMin
          in: query
          required: false
          schema: { type: integer }
          description: Minimum rank (inclusive).
        - name: rankMax
          in: query
          required: false
          schema: { type: integer }
          description: Maximum rank (inclusive).
        - name: commonLocation
          in: query
          required: false
          schema:
            type: string
          description: Filter by exact common location string (matches any element of commonLocations[]). Comma-separated for multiple.
          example: "Hyrule Field"
        - name: itemDrop
          in: query
          required: false
          schema:
            type: string
          description: Filter by item drop name (exact match) or drop id. Comma-separated for multiple (matches if any listed drop is present).
          example: "Bokoblin Horn"
      responses:
        '200':
          description: Array of monster objects (filtered if query params provided)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Monster' }
  /monsters/{id}:
    get:
      tags: [Monsters]
      summary: Get monster by ID
      operationId: getMonsterById
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Single monster object
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Monster' }

  /shields:
    get:
      tags: [Shields]
      summary: List shields (with optional filters)
      operationId: getShields
      description: |
        Returns shields. Optionally narrow by durability and parry power ranges, and/or availability.
        Multiple availability values use a comma-separated string (matches if any listed value appears in availabilities[]).
      parameters:
        - name: durabilityMin
          in: query
          required: false
          schema: { type: integer }
          description: Minimum durability (inclusive).
        - name: durabilityMax
          in: query
          required: false
          schema: { type: integer }
          description: Maximum durability (inclusive).
        - name: parryPowerMin
          in: query
          required: false
          schema: { type: integer }
          description: Minimum parry power (inclusive).
        - name: parryPowerMax
          in: query
          required: false
          schema: { type: integer }
          description: Maximum parry power (inclusive).
        - name: availability
          in: query
          required: false
          schema:
            type: string
          description: Filter by exact availability string (matches any element of availabilities[]). Comma-separated for multiple.
          example: "Hyrule Field"
      responses:
        '200':
          description: Array of shield objects (filtered if query params provided)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Shield' }
  /shields/{id}:
    get:
      tags: [Shields]
      summary: Get shield by ID
      operationId: getShieldById
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Single shield object
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Shield' }

  /weapons:
    get:
      tags: [Weapons]
      summary: List all weapons
      operationId: getWeapons
      responses:
        '200':
          description: Array of weapon objects
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Weapon' }
  /weapons/{id}:
    get:
      tags: [Weapons]
      summary: Get weapon by ID
      operationId: getWeaponById
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Single weapon object
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Weapon' }

  /bows:
    get:
      tags: [Bows]
      summary: List bows (with optional filters)
      operationId: getBows
      description: |
        Returns bows. Optionally narrow by numeric ranges (attack power, durability, range, multiple arrows),
        a Quick Shot boolean, and/or availability. Multiple availability values use a comma-separated string.
      parameters:
        - name: attackPowerMin
          in: query
          required: false
          schema: { type: integer }
          description: Minimum attack power value (inclusive).
        - name: attackPowerMax
          in: query
          required: false
          schema: { type: integer }
          description: Maximum attack power value (inclusive).
        - name: durabilityMin
          in: query
          required: false
          schema: { type: integer }
          description: Minimum durability value (inclusive).
        - name: durabilityMax
          in: query
          required: false
          schema: { type: integer }
          description: Maximum durability value (inclusive).
        - name: rangeMin
          in: query
          required: false
          schema: { type: integer }
          description: Minimum range value (inclusive).
        - name: rangeMax
          in: query
          required: false
          schema: { type: integer }
          description: Maximum range value (inclusive).
        - name: multipleArrowsMin
          in: query
          required: false
          schema: { type: integer }
          description: Minimum multiple arrows value (inclusive).
        - name: multipleArrowsMax
          in: query
          required: false
          schema: { type: integer }
          description: Maximum multiple arrows value (inclusive).
        - name: quickShot
          in: query
          required: false
          schema: { type: boolean }
          description: Filter by Quick Shot (true/false). Omit to include all.
        - name: availability
          in: query
          required: false
          schema:
            type: string
          description: Filter by exact availability string (matches any element of availabilities[]). Comma-separated for multiple.
          example: "Hyrule Field"
      responses:
        '200':
          description: Array of bow objects (filtered if query params provided)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Bow' }
  /bows/{id}:
    get:
      tags: [Bows]
      summary: Get bow by ID
      operationId: getBowById
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Single bow object
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Bow' }

  /animals:
    get:
      tags: [Animals]
      summary: List animals (with optional filters)
      operationId: getAnimals
      description: Returns animals. Optionally filter by unique cooking effect, common location, and/or recoverable material. Multiple values per parameter are supported as comma-separated strings.
      parameters:
        - name: uniqueCookingEffect
          in: query
          required: false
          schema:
            type: string
          description: Filter by unique cooking effect (e.g. "Attack Up", "Defense Up"). Comma-separated for multiple (e.g. "Attack Up,Defense Up").
          example: "Attack Up"
        - name: commonLocation
          in: query
          required: false
          schema:
            type: string
          description: Filter by common location name (e.g. "Hyrule Field", "Hebra Mountains"). Comma-separated for multiple.
          example: "Hyrule Field"
        - name: recoverableMaterial
          in: query
          required: false
          schema:
            type: string
          description: Filter by recoverable material name or material id. Comma-separated for multiple (e.g. "Raw Meat,Raw Prime Meat" or "40,38").
          example: "Raw Meat"
      responses:
        '200':
          description: Array of animal objects (filtered if query params provided)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Animal' }
  /animals/{id}:
    get:
      tags: [Animals]
      summary: Get animal by ID
      operationId: getAnimalById
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Single animal object
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Animal' }

  /other:
    get:
      tags: [Other]
      summary: List all other items
      operationId: getOther
      description: Miscellaneous in-game items (rupees, key items, bridles, saddles, compendium categories, etc.).
      responses:
        '200':
          description: Array of other item objects
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/OtherItem' }
  /other/{id}:
    get:
      tags: [Other]
      summary: Get other item by ID
      operationId: getOtherById
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Single other item object
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OtherItem' }

components:
  schemas:
    Material:
      type: object
      description: Game material (e.g. fruit, insect, ore)
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string }
        cssClassName: { type: string }
        description: { type: string }
        sellPrice: { type: number }
        hpRecovery: { type: number }
        category: { type: object }
        availabilities: { type: array, items: { type: string } }
        recipes: { type: array }
    Recipe:
      type: object
      description: Cooking recipe
      properties:
        id: { type: integer }
        name: { type: string }
        cssClassName: { type: string }
        materials: { type: array }
        effects: { type: object }
    Armor:
      type: object
      description: Armor piece
      properties:
        id: { type: integer }
        name: { type: string }
        cssClassName: { type: string }
        defense: { type: number }
        set: { type: string }
        upgradeLevel: { type: integer }
    Food:
      type: object
      description: Cooked food item
      properties:
        id: { type: integer }
        name: { type: string }
        cssClassName: { type: string }
        effects: { type: object }
    Monster:
      type: object
      description: Enemy or creature
      properties:
        id: { type: integer }
        name: { type: string }
        cssClassName: { type: string }
        description: { type: string }
        itemDrops: { type: array }
    Shield:
      type: object
      description: Shield equipment
      properties:
        id: { type: integer }
        name: { type: string }
        cssClassName: { type: string }
        defense: { type: number }
    Weapon:
      type: object
      description: Weapon equipment
      properties:
        id: { type: integer }
        name: { type: string }
        cssClassName: { type: string }
        attack: { type: number }
    Bow:
      type: object
      description: Bow equipment
      properties:
        id: { type: integer }
        name: { type: string }
        cssClassName: { type: string }
        attack: { type: number }
    Animal:
      type: object
      description: Wildlife (non-enemy)
      properties:
        id: { type: integer }
        name: { type: string }
        cssClassName: { type: string }
        description: { type: string }
        commonLocations: { type: array, items: { type: string } }
        recoverableMaterials: { type: array, items: { type: object, properties: { id: { type: integer }, name: { type: string } } } }
        uniqueCookingEffects: { type: string }
    OtherItem:
      type: object
      description: Miscellaneous item (rupees, key items, etc.)
      properties:
        id: { type: integer }
        name: { type: string }
        cssClassName: { type: string }
        description: { type: string }